refactor(tests): extract shared MCP harness into tests/common/mod.rs
